fixes #51: Refactor swin_result_parser function to add all result to the return form
fixes #51: Change model_utilitary_function to model_request module and update app.py with new imports and function calls

fixes #51: fix bug with request import
