+++ model_inference/inference.py
@@ -1,6 +1,68 @@
 import numpy as np
+import io
+import base64
+from PIL import Image
 from custom_exceptions import ProcessInferenceResultError
 
+async def image_slicing(image_bytes: bytes, result_json: dict) -> list:
+    """
+    This function takes the image bytes and the result_json from the model and
+    returns a list of cropped images.
+    The result_json is expected to be in the following format:
+    {
+        "boxes": [
+            {
+                "box": {
+                    "topX": 0.0,
+                    "topY": 0.0,
+                    "bottomX": 0.0,
+                    "bottomY": 0.0
+                },
+                "label": "string",
+                "score": 0.0
+            }
+        ],
+    }
+    """
+    boxes = result_json[0]['boxes']
+    image_io_byte = io.BytesIO(base64.b64decode(image_bytes))
+    image_io_byte.seek(0)
+    image = Image.open(image_io_byte)
+
+    format = image.format
+
+    cropped_images = [bytes(0) for _ in boxes]
+
+    for i, box in enumerate(boxes):
+        topX = int(box['box']['topX'] * image.width)
+        topY = int(box['box']['topY'] * image.height)
+        bottomX = int(box['box']['bottomX'] * image.width)
+        bottomY = int(box['box']['bottomY'] * image.height)
+
+        img = image.crop((topX, topY, bottomX, bottomY))
+
+        buffered = io.BytesIO()
+        img.save(buffered, format)
+
+        cropped_images[i] = base64.b64encode(buffered.getvalue()) #.decode("utf8")
+    
+    return cropped_images
+
+async def swin_result_parser(img_box:dict, results: dict) -> list:
+    """
+    Args:
+        img_box (dict): The image box containing the bounding boxes and labels.
+        results (dict): The results from the model containing the detected seeds.
+
+    Returns:
+        list: The updated image box with modified labels and scores.
+    """
+    for i, result in enumerate(results):
+        img_box[0]['boxes'][i]['label'] = result[0].get("label")
+        img_box[0]['boxes'][i]['score'] = result[0].get("score")
+        img_box[0]['boxes'][i]["all_result"] = [d for d in result]
+    
+    return img_box
 
 async def process_inference_results(data, imageDims):
     """
